contact apis,auth middleware

+++ learn-backend/controllers/contactController.js
@@ -2,16 +2,16 @@ const asyncHandler = require("express-async-handler")
 const Contact = require("../models/contactModel")
 //@decription : get all contact
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find({});
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@decription : create contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     // console.log(req.body)
     const { name, email, phone } = req.body;
@@ -20,14 +20,14 @@ const createContact = asyncHandler(async (req, res) => {
         throw new Error("All the fields are required")
     }
     const contacts = await Contact.create({
-        name, email, phone
+        name, email, phone, user_id: req.user.id
     })
     res.status(201).json(contacts);
 })
 
 //@decription : get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
     if (!contact) {
@@ -39,12 +39,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@decription : update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
     if (!contact) {
         res.status(404);
         throw new Error("No contact found")
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403)
+        throw new Error("user dont have permission to update contact");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -56,16 +61,19 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@decription : update contact
 //@route delete /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("No contact found");
-    } else {
-        await Contact.deleteOne({ _id: req.params.id });
-        return res.status(200).json({ message: "Contact deleted successfully" });
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403)
+        throw new Error("user dont have permission to delete contact");
+    }
+    await Contact.deleteOne({ _id: req.params.id });
+    return res.status(200).json({ message: "Contact deleted successfully" });
 });
 
 module.exports = { getContacts, createContact, getContact, updateContact, deleteContact };
